Revert "string-mutableandimmutable"
